update docs and move logout handler
