+++ main.go
@@ -28,8 +28,8 @@ func main() {
 	router.POST("/api/v1/account/sign-up", api.SignUpHandler)
 	router.POST("/api/v1/account/login", api.LoginHandler)
 	router.POST("/api/v1/account/security-questions", middleware.JWTAuthMiddleware(), api.SecurityQuestionsHandler)
+	router.POST("/api/v1/account/logout", api.Logout)
 
-	router.POST("/api/v1/jwt/logout", jwt.Logout)
 	router.POST("/api/v1/jwt/forgot-password", jwt.ForgotPasswordHandler)
 	router.GET("/api/v1/jwt/auth-check", jwt.AuthCheckHandler)
 	router.GET("/api/v1/example/jwt", middleware.JWTAuthMiddleware(), api.ExampleAuthEndpoint)
